agregar controladores de images
