+++ routes/images.routes.js
@@ -1,30 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const path = require('path');
-const multer = require('multer');
-const fs = require('fs');
+const {uploadImages} = require('../controllers/image.controller')
 
-
-const storage = multer.diskStorage({
-    destination: path.join(__dirname, '../public/uploads'),
-    filename:  (req, file, cb) => {
-        cb(null, file.originalname);
-    }
-})
-const uploadImage = multer({
-    storage,
-    limits: {fileSize: 1000000}
-}).single('image');
-
-router.post('/upload', (req, res) => {
-    uploadImage(req, res, (err) => {
-        if (err) {
-            err.message = 'The file is so heavy for my service';
-            return res.send(err);
-        }
-        console.log(req.file);
-        res.send('uploaded');
-    });
-});
+router.post('/upload', uploadImages);
 
 module.exports = router;
